ADDED A DYNAMIC-PAGE FOR EACH MEMBER
Clicking on a members snippet redirects to a new page with the member information.
In order to load dinamically the member info into the template 'ficha-miembro.html' the click event stores some parameters into the sessionStorage. These parameters are used as a help to render the content into new landing page 'ficha-miembros.html'.